feat: 쿠폰 발급 요청 시 이메일 입력 및 중복 검사 (#15)
* feat: 쿠폰 발급 시 이메일 입력

* feat: 쿠폰 발급 시 이메일 입력 형식 검증

* test: RestAssured 호출 fixture 분리

* test: DatabaseCleanup 추가

* feat: 당일 사용된 이메일로 쿠폰 발급 요청 시 Bad Request 응답

* feat: 쿠폰 수량 관리 redis set으로 변경

* feat: Redis Transaction 적용 위한 `TransactionalRedisOperations` 추가

* feat: 쿠폰 수량 조회 및 발급 한 transaction에서 처리

+++ src/main/java/com/coupop/fcfscoupon/model/CouponIssuePolicy.java
@@ -16,7 +16,7 @@ public class CouponIssuePolicy {
         return issuedCount >= LIMIT;
     }
 
-    public static boolean isCouponOverIssued(int issuedCount) {
+    public static boolean isCouponOverIssued(Long issuedCount) {
         return issuedCount > LIMIT;
     }
